Made existing code compile with `-Wall -Werror -Wextra`.

+++ format_disk_as_pantryfs.c
@@ -120,7 +120,7 @@ int main(int argc, char *argv[])
 
 	len = PFS_BLOCK_SIZE - sizeof(struct pantryfs_dir_entry);
 	ret = write(fd, zeroes, len);
-	passert(ret == len, "Pad to end of root dentries");
+	passert(ret == (ssize_t)len, "Pad to end of root dentries");
 
 	strncpy(buf, hello_contents, sizeof(buf));
 	ret = write(fd, buf, sizeof(buf));
